+++ dotman.sh
@@ -3,11 +3,36 @@
 export BASE_DIR
 BASE_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-source lib/packages.sh
-source lib/links.sh
-source lib/scripts.sh
 source lib/paths.sh
 source lib/execute.sh
+
+##################################################
+#                                                #  
+#  [dotman.sh] - v0.0.1                          #  
+#                                                #  
+#        The declarative dotfiles manager        #  
+#                                                #  
+#                                                #  
+#                               Made by KDesp73  #  
+#                                                #  
+##################################################
+
+##################################################
+#                                                #  
+#  Structure:                                    #  
+#                                                #  
+#  dotfiles/                                     #  
+#          |--lib/ # The library files           #  
+#          |                                     #  
+#          |--scripts/ # Your custom scripts     #  
+#          |                                     #  
+#          |--dotman.sh # The entry point        #  
+#          |                                     #  
+#          |--<all your dotfiles>                #  
+#                                                #  
+##################################################
+
+# VVV Edit below VVV
 
 pkgs=(
     git
@@ -27,10 +52,4 @@ links[".zshrc"]="$HOME"
 links["nvim"]="$CONFIG"
 # more...
 
-run () {
-    install_packages pkgs
-    run_scripts scripts
-    linker links #<optional prefix>
-}
-
-execute run "$@"
+execute pkgs scripts links "$@"