+++ lib/links.sh
@@ -23,18 +23,43 @@ linker() {
     local -n hashmap=$1
     local prefix=$2
 
+    local rc=0
     for key in "${!hashmap[@]}"; do
         color green "Linking $key..."
 
         if ! file_exists "$prefix$key" ; then
-            ERRO "File '$prefix$key' does not exist" 
-            return 1
+            ERRO "No such file or directory: $prefix$key" 
+            rc=1
+            continue
         fi
 
         if ! link "$prefix"/"$key" "${hashmap[$key]}"; then
             ERRO "Link for ${hashmap[$key]}/$key already exists"
-            return 1
+            rc=1
         fi
     done
-    return 0
+    return $rc
+}
+
+remove_links() {
+    if ! read_yes_no "This will remove all symlinks. Are you sure you want to continue?" ; then
+        return 1
+    fi
+
+    local -n hashmap=$1
+    local prefix=$2
+
+    local rc=0
+    for key in "${!hashmap[@]}"; do
+        color red "Removing link $key..."
+
+        if ! file_exists "$prefix$key" ; then
+            ERRO "File '$prefix$key' does not exist" 
+            rc=1
+            continue
+        fi
+
+        rm "${hashmap[$key]}"
+    done
+    return $rc
 }